pyproject.toml
[tool.poetry]
name = "request-checker"
version = "0.1.0"
description = ""
authors = ["jsik94 <ykmgameha@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.112.2"
uvicorn = "^0.30.6"
pandas = "^2.2.2"
python-dotenv = "^1.0.1"
jinja2 = "^3.1.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
